requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==3.2.13
django-cors-headers==3.14.0
django-extensions==3.2.1
django-seed==0.3.1
djangorestframework==3.14.0
executing==1.2.0
Faker==18.9.0
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toposort==1.10
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
